Shift the follow and unfollow routes from currUser to user file
All the routes in the user file are related to the currently logged user. In
Future the route path will be changed. But follow and unfollow is the user actions
not the personal user actions.

Hence, to bring the routes in proper place shift the follow and unfollow routes
to users file.

Since, this is testing phase there is no side effects anywhere.