Update the follow and unfollow routes
Follow Route: Initially currUserDetails and toFollowUserId both was expected to be
given from frontend, but after the addition of jwt authentication we have the details
of currUser in the request object.

Change the way the currUserDetails is accessed, and change the variable name for better
understanding of the code

UnFollow Route: The way both the user id was accessed prior was the same here. Did the
same updation here for accessing the currUserDetails.
Fixed the bug in the logic for unfollowing the user